+++ mcstools/__init__.py
@@ -2,7 +2,7 @@ from data_path_handler import FilenameBuilder
 from loader import L1BLoader, L2Loader
 from mcsfile import L1BFile, L2File
 from preprocess.data_pipeline import L1BDataPipeline
-from preprocess.l1b_limb import L1BStandardInTrack
+from preprocess.l1b import L1BOnPlanetInTrack, L1BStandardInTrack
 from reader import L1BReader, L2Reader
 from util import geom, log, mars_time, time
 
@@ -14,6 +14,7 @@ __all__ = [
     "L2File",
     "L1BDataPipeline",
     "L1BStandardInTrack",
+    "L1BOnPlanetInTrack",
     "L1BReader",
     "L2Reader",
     "geom",
